requirements/base.txt
fastapi==0.133.1
uvicorn==0.41.0
sqlmodel==0.0.37
SQLAlchemy==2.0.47
alembic==1.18.4
psycopg2-binary==2.9.11
python-dotenv==1.2.1
httpx==0.28.1
openai==2.24.0
requests==2.32.5
funkybob==2023.12.0
numpy==2.4.2
scikit-learn==1.8.0
sentence-transformers==5.2.3
